refactor: extract reusable menu widgets
